+++ kibble/utils/launch.go
@@ -1,0 +1,26 @@
+package utils
+
+import (
+	"fmt"
+	"os/exec"
+	"runtime"
+)
+
+// LaunchBrowser will open a browser
+func LaunchBrowser(url string) {
+	var err error
+
+	switch runtime.GOOS {
+	case "linux":
+		err = exec.Command("xdg-open", url).Start()
+	case "windows":
+		err = exec.Command("rundll32", "url.dll,FileProtocolHandler", url).Start()
+	case "darwin":
+		err = exec.Command("open", url).Start()
+	default:
+		err = fmt.Errorf("unsupported platform")
+	}
+	if err != nil {
+		log.Fatal(err)
+	}
+}